feat: Ensure sufficient color contrast for the class `text-light-emphasis`.
The logo color did not contrast the header background enough for accessibility standards. I changed the logo element class from `logo-color` to `text-light-emphasis`  (want to use the Bootstrap5 classes as much as possible).  This class uses the variable `--bs-light-text-emphasis`. I made the color lighter in dark mode, to increase the contrast.