feat/fix: Changes to style, CSS, classes.
- feat: Use Bootstrap5 classes instead of CSS in sidebar.

- feat: Remove unnecessary classes.

- feat: remove inline style.

- fix: Display link with active-styling when active.

When the page "/projects/visualize" was active, the lik in the sidebar was not changing colors like the other sidebar links (when they were active). This was because this link's href-property (in 'links-tsx') had an extra "/" at the end, which caused it to not match the result of `usePathname()` when active. This link was therefore never given the class name 'nav-link active''.

- feat: Change BS5 colors.